Moved use-logic to handler-functions.
This is similar to what I need to do for the take/drop logic,
to allow me to define z80 macros for objects.

Moving the logic into per-object handlers means our handler-code
doesn't need to have special cases built in.

To test this out we've replaced the "TORCH" with two items:

* Torch
  * A normal torch, unlit.
* Torch-Lit
  * A torch which will save you from the grue.