Further evolution of the bios_delay function.
This function has been a lot of fun to write!  Initially I wrote
an interrupt handling routine to count frames, this would get called
every 50ms, unfortunately by taking over the IM2 handler the BASIC
keyboard input routine was ignored.

I could have fixed this via RST 56, but instead I read the ROM
disassembly and realized that the ROM also maintains a frame counter..

So the second attempt was to use the frame-counter the ROM updated,
removing my own handler.

This is the third attempt!  Recognizing that there is an interrupt
handler I can just call `halt`.  The processor will wake up after
the next interrupt.

Wait for 5?  That's 5/50ms.  Which is a good display.