Allow the torch to be toggled on/off cleanly
This updates the code to avoid having two torch-objects which
are swapped (TORCH + TORCH-LIT).  Instead the state of the torch
is determined by the STATE bye in the object-table.

This simplifies some of the code, at the cost of having to swap
descriptions about when the state-changes.

This will close #30 once the trapdoor code is updated similarly.