The trapdoor has three logical states.
We have added a logical "invisible" state to the trapdoor.  All of
our objects are already allowed to be invisible, but they tend to
be things that can't be interacted with, or fake-invisible so that
you can access them if you know about them.

The trapdoor must be really invisible though, so it can only be
opened/closed once discovered.

Add the third state, and use it appropriately.  This closes #33.