Ensure that the first code executed is our decryption routine
This pull-request closes #37, by making sure the first code we
run in our binary is our decryption code (if that is enabled).

Previously we ran some code:

        call foo
        [decrypt code]
        [start of encrypted code]
        foo:
        [end of encrypted code]

If we called a routine that was after the decryption code we'd
be running gibberish.  This seemed to work by accident in some
systems, but was obviously broken.