Generate encrypted versions of the ZX Spectrum game
The recent issue with broken encryption, reported in #37, might have
been more easily visible if I'd generated encrypted versions of the
ZX Spectrum image.

Generating an encrypted ZX Spectrum binary is largely the same as
the process for the CP/M binary - with one complication:

* Generate the build, with the DEFINE set.
* Post-process the generated build, to XOR the contents.

The complication is that the .TAP file generated for Spectrums contains
checksums, so we needed to update the `encrypt.c` helper to add support
for that (optionally).

I've updated the `release` target of the Makefile to consider ordering
too, such that we generate "naked" and "encrypted" releases of both
our versions.

All together this should close #38.