Prepare for word-wrapping
This pull-request, once complete, will implement a decent amount of
line-wrapping.

There are two distinct things that need to happen for word-wrapping
to be implemented:

1.  Remove the hardcoded 0x0a, 0x0d from the text-strings.  These were
    added to provide "natural" line-breaks when developing for CP/M.

2.  Update the output routine to start a new line on a word-boundary.

Beyond that we need to allow a little flexibility on the configuration,
this WIP-version has two hardcoded widths which isn't ideal but will
suffice for a starting-point.