Correctly terminate the game cleanly on death/quit
Previously the code would print some random gibberish and exit
in an unclean fashion after the user declined the opportunity
to play again.

looking at this closely it was because I used the wrong offset of
the "quitting" message to print - so random binary junk would get
displayed before a $ character was encountered by chance, and terminated
the display.

Correctly point to the "terminating" message, and exit cleanly to
BASIC/CP/M.