Add a simple per-object table for custom actions.
This commit adds a simple table for per-item actions, and moves
the "READ" function into the entry for the book.

We don't yet invoke the handler, but we do successfully look for
a) the object, and b) the custom verbs for it.  If they're not
present we continue - otherwise we show a temporary debugging
message to confirm an action was found.