Simplify our input handling, and implement trapdoor/torch
In the past we removed " THE " and " ON " from input, to avoid
issues with "TAKE THE RUG" and "TURN ON TORCH".  Now we've updated
so that our routine to fetch the "second" word always takes the
*last* word:

* TAKE THE TORCH -> TORCH
  * Not "THE".

This means we can have an abitrary prefix for custom actions,
and that means we don't need to care about the extra ON/THE.

Also I've implemented OPEN/CLOSE trapdoor and TURN ON/OFF/LIGHT
the torch.  These "just" do a toggle, but that's okay for now.

We only have one custom use function left, which refers to the
generator.  We should probably change that to "start".